refactor IUser interface from UserCard component to common/types
