+++ src/common/types.ts
@@ -1,5 +1,6 @@
 import { PreloadedQuery, UseQueryLoaderLoadQueryOptions } from "react-relay";
 import { GraphQLTaggedNode, OperationType, VariablesOf } from "relay-runtime";
+import { Maybe } from "src/generated/graphql";
 
 export type PagePropsWithQuery<T extends OperationType> = {
     queryReference: PreloadedQuery<T>;
@@ -11,3 +12,11 @@ export type QueryInfo = {
     preloadedQuery: PreloadedQuery<OperationType>;
     queryObject: any;
 };
+
+export interface IUser {
+    name?: string;
+    email?: string;
+    isActive?: boolean;
+    avatar?: Maybe<string>;
+    userId?: string;
+}
